* Removed unused files. * Improved directory structure. * Upgraded the dependencies. * Fixed migrations. * Updated documentation. * Fixed some bugs.
