+++ config/routes.php
@@ -10,11 +10,4 @@
  */
 declare(strict_types=1);
 
-use Webify\Tools\Rector\Rector;
-
-return (new Rector())->initialize(
-    [
-        __DIR__ . '/src',
-        __DIR__ . '/test',
-    ]
-);
+return [];
