--- src/Infrastructure/Persistence/Person/Query/PersonQuery.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Webify\User\Infrastructure\Persistence\Person\Query;
-
-final class PersonQuery
-{
-}
