--- templates/person/create.php
@@ -1,12 +1,0 @@
-<?php
-
-use Webify\User\Infrastructure\Presentation\Admin\Form\Person\CreateFormPerson;
-
-/**
- * @var CreateFormPerson $form
- */
-?>
-
-<header class="">
-    <h1 class="">Create Person</h1>
-</header>
